+++ src/MenuCharacter.h
@@ -81,7 +81,7 @@ private:
 
 	std::vector<int*> base_stats;
 	std::vector<int*> base_stats_add;
-	std::vector< std::vector<int>* > base_bonus;
+	std::vector< std::vector<float>* > base_bonus;
 
 	int name_max_width;
